requirements.txt
#requirements.txt
python-dotenv
discord.py
numpy
motor
